More code dedupe + possible memory leak prevention
