+++ src/media/VideoStream.ts
@@ -1,4 +1,3 @@
-import { setTimeout as setTimeoutPromise } from "timers/promises";
 import { MediaUdp } from "../client/voice/MediaUdp.js";
 import { combineLoHi } from "./utils.js";
 import { BaseMediaStream } from "./BaseMediaStream.js";
@@ -9,10 +8,8 @@ export class VideoStream extends BaseMediaStream {
     public count: number;
     public sleepTime: number;
     public startTime?: number;
-    public syncStream?: BaseMediaStream;
 
     private noSleep: boolean;
-    private _syncTolerance: number = 0;
 
     constructor(udp: MediaUdp, fps: number = 30, noSleep = false) {
         super({ objectMode: true });
@@ -26,35 +23,17 @@ export class VideoStream extends BaseMediaStream {
         this.sleepTime = time;
     }
 
-    get syncTolerance() {
-        return this._syncTolerance;
-    }
-
-    set syncTolerance(n: number) {
-        if (n < 0)
-            return;
-        this._syncTolerance = n;
-    }
-
     async _write(frame: Packet, encoding: BufferEncoding, callback: (error?: Error | null) => void) {
         this.count++;
         if (!this.startTime)
             this.startTime = performance.now();
 
-        // We are ahead, wait for the other stream to catch up
-        while (
-            this.syncStream &&
-            !this.syncStream.writableEnded &&
-            this.syncStream.pts !== undefined &&
-            this._pts !== undefined &&
-            this._pts - this.syncStream.pts > this._syncTolerance
-        )
-            await setTimeoutPromise(1);
+        await this._waitForOtherStream();
 
         const { data, ptshi, pts, time_base_num, time_base_den } = frame;
         this.udp.sendVideoFrame(Buffer.from(data));
         if (ptshi !== undefined && pts !== undefined && time_base_num !== undefined && time_base_den !== undefined)
-            this._pts = combineLoHi(ptshi, pts) / time_base_den * time_base_num;
+            this.pts = combineLoHi(ptshi, pts) / time_base_den * time_base_num;
 
         const next = ( (this.count + 1) * this.sleepTime) - (performance.now() - this.startTime);
